Lava Glow and Movement
I was able to make the lava glow and move it's textures to make it flow. You can change the speed and direction of the lava. also you can make it brighter if you want. I did have to download a package manager in order for the glow can work. It's in the unity Register and it's called post processing. I had to add an empty object to the scene and add a post-process volume on it. that's where one can control the glow. I also had to add a post-process layer to the camera. we don't change anything there other than setting the layer to everything. I made the empty object a prefab so we can add it any scene where there is lava.